Thorify the CLI command
This sets up the CLI to be a Thor tool. We chose Thor mostly because we
eventually need to hook into Rails generators, which are based on Thor.
So we figured, why not just use Thor for everything! ヽ(。_°)ノ

We define `exit_on_failure?` per the Thor wiki. This is necessary to
ensure the exe exit status is properly set to non-zero on error.

See https://github.com/erikhuda/thor/wiki/Making-An-Executable