Refactor the helpers to work with classes
This refactors our raw common setup helpers into something that will
work better with the CLI suite. The original file was called by
individual scripts so there was no concern about polluting the global
namespace.

This switches to encapsulate everything within a module. We make the
helper methods available via module functions as they do not rely on
state and that makes them private implementation details when included.
Similarly we scope the constant and make it available as a helper.

As everything is now namespaced we rename the file/module to just
`Helpers`; as the intent is clear within this CLI suite.