Fix ordering issue with puma-dev cert checks
In the case puma-dev is never setup the script will skip everything and
not inform the user. This move the cert check to the setup step because
we cannot register it with the OS unless it exists. After that point we
know the cert exists so we do not need to check it again.