Update Thor requirement to support Rails 6.1+
We're currently working on upgrading Captain to Rails 6.1, which
requires version 1.0 of Thor. This gem, however, had the Thor version
pinned to 0.20. To support the Captain upgrade to Rails 6.1 we're
upgrading the Thor dependency to 1.0. Looking over the Thor CHANGELOG
there are no major breaking changes that we should need to worry about.